DEV-1897 Change Dockerfile
In the Dockerfile, a new user and group is created so that it doesn't run as root.
However, this application needs access to certain folders on the host.

Make the ID of the user and group adjustable via ARG parameters so that the
container can be run as a local user on the host.