DEV-2067 Use CP ID from the watchfolder message
The CP ID was fetched from the VIAA metadata in the sidecar. This is
information the CPs themselves deliver. Meaning, that this information is
not necessarily correct as we are owners of the CP IDs and we assign such
a CP ID to the CP.

The watchfolder message does contain the correct information as this comes
from internal configuration. Do note that it is called flow id in the
watchfolder message.