DEV-2081 Copy essence when creating the meemoo SIP
The meemoo SIP bag folder gets removed after successfully zipping that bag.
That zipped bag gets removed by the `sipin-delete-service` after a successful
transfer of the zipped bag. The essence, however should remain on the FTP
until after a successful ingest of the essence in the MAM. Copying the essence
to the bag folder instead of moving it ensures that the essence stays on
the FTP as the `sipin-delete-service` deletes the original essence and sidecar
after successful ingest.