DEV-2098 Calculate original name from sidecar
According to the VIAA metadatamodel spec, the node containing the original
name is described by the XPATH: `dc_identifier_localids/Bestandsnaam`. However,
those localids are a kind of free for all, meaning that `bestandsnaam` is also
possible. Furthermore, sometimes `dc_source` is also used to describe the original
name.

To calculate the original name, first of, check in the sidecar metadata
in order of existence:
 -VIAA/dc_identifiers_localids/Bestandsnaam
 -VIAA/dc_identifiers_localids/bestandsnaam
 -VIAA/dc_source

If none are available, use the filename of the essence.