DEV-2098 Use fixed mimetypes map
It used the Python module mimetypes, which guessed the mimetype based
on the extension of the file. Python loads in this mimetype map information
from certain files on the OS, e.g. /etc/mime.types. However, this gave
non-deterministic results over different platforms. As the extensions of the
incoming files are a fixed list, we can use a hardcoded mimetype map instead.