DEV-2103 Don't compress zip file
The higher-level function `shutil.make_archive()` seems to compress the zip
file. As this is mostly audiovisual material, compression is not desired.
However, compressing the zip file takes quite some time. In order to win some
time, we do not compress the zip file.

The function `shutil.make_archive()` doesn't seem to have a compression-level
param. So we switch to the lower-level Zipfile instead.