DEV-2121 Get CP name from org api
The CP name is taken from the watchfolder message. This is information
that is configured in the watchfolders. However, the more "correct"
information can be requested from the org API, that serves that information
from Teamleader.

Get the CP name from the org API instead of from the watchfolder message.